switch to float (doesn't increase speed)

+++ CudaBEM/Funcs.cuh
@@ -2,13 +2,13 @@
 
 #include "Point.h"
 
-__host__ __device__ double F1(Vector x, Vector y, double p)
+__host__ __device__ float F1(Vector x, Vector y, float p)
 {
-	return p / (4 * 3.14159265358979323846 * (x - y).Norm());
+	return p / (4 * 3.14159265358979323846f * (x - y).Norm());
 }
 
-__host__ __device__ double F2(Vector x, Vector y, Vector n, double q)
+__host__ __device__ float F2(Vector x, Vector y, Vector n, float q)
 {
-	double norm = (x - y).Norm();
-	return ((n * q) * (x - y)) / (4 * 3.14159265358979323846 * norm * norm * norm);
+	float norm = (x - y).Norm();
+	return ((n * q) * (x - y)) / (4 * 3.14159265358979323846f * norm * norm * norm);
 }
